DT-727 Split functional tests and parameterize for Kotlin/Java
