+++ src/main/kotlin/uk/gov/justice/digital/hmpps/gradle/configmanagers/BaseConfigManager.kt
@@ -32,7 +32,7 @@ class BaseConfigManager(override val project: Project) : ConfigManager {
 
   private fun addDependencies() {
     project.dependencies.add("implementation", "com.fasterxml.jackson.module:jackson-module-kotlin")
-    project.dependencies.add("implementation", "com.google.guava:guava:29.0-jre") // This is only required because the version pulled in a as transitive dependency has CVE vulnerabilities
+    project.dependencies.add("implementation", "com.google.guava:guava:29.0-jre") // This is only required because the version pulled in as a transitive dependency has CVE vulnerabilities
   }
 
 }
