+++ src/test/kotlin/uk/gov/justice/digital/hmpps/gradle/functional/pluginmanagers/SpringBootPluginManagerTest.kt
@@ -1,0 +1,127 @@
+package uk.gov.justice.digital.hmpps.gradle.functional.pluginmanagers
+
+import org.assertj.core.api.Assertions.assertThat
+import org.gradle.internal.impldep.org.codehaus.plexus.util.FileUtils
+import org.gradle.testkit.runner.TaskOutcome
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.io.TempDir
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
+import org.junit.jupiter.params.provider.MethodSource
+import uk.gov.justice.digital.hmpps.gradle.functional.ProjectDetails
+import uk.gov.justice.digital.hmpps.gradle.functional.buildProject
+import uk.gov.justice.digital.hmpps.gradle.functional.buildProjectAndFail
+import uk.gov.justice.digital.hmpps.gradle.functional.findJar
+import uk.gov.justice.digital.hmpps.gradle.functional.getDependencyVersion
+import uk.gov.justice.digital.hmpps.gradle.functional.javaProjectDetails
+import uk.gov.justice.digital.hmpps.gradle.functional.kotlinProjectDetails
+import uk.gov.justice.digital.hmpps.gradle.functional.makeProject
+import java.io.File
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+import java.util.jar.JarFile
+
+class SpringBootPluginManagerBuildTest {
+
+  @AfterEach
+  fun `Delete project`() {
+    FileUtils.cleanDirectory(projectDir)
+  }
+
+  companion object {
+
+    @TempDir
+    @JvmStatic
+    lateinit var projectDir: File
+
+    @JvmStatic
+    fun defaultProjectDetails() = listOf(
+        Arguments.of(javaProjectDetails(projectDir)),
+        Arguments.of(kotlinProjectDetails(projectDir))
+    )
+
+    @JvmStatic
+    fun projectDetailsWithJunit4Tests() = listOf(
+        Arguments.of(javaProjectDetails(projectDir).copy(testClass = javaJunit4Test())),
+        Arguments.of(kotlinProjectDetails(projectDir).copy(testClass = kotlinJunit4Test()))
+    )
+  }
+
+  @ParameterizedTest
+  @MethodSource("defaultProjectDetails")
+  fun `Spring dependency versions are defaulted from the dependency management plugin`(projectDetails: ProjectDetails) {
+    makeProject(projectDetails)
+
+    val webVersion = getDependencyVersion(projectDir, "spring-boot-starter-web")
+    val actuatorVersion = getDependencyVersion(projectDir, "spring-boot-starter-actuator")
+
+    assertThat(webVersion).isEqualTo(actuatorVersion)
+  }
+
+  @ParameterizedTest
+  @MethodSource("defaultProjectDetails")
+  fun `Manifest file contains project name and version`(projectDetails: ProjectDetails) {
+    makeProject(projectDetails)
+
+    val result = buildProject(projectDir, "bootJar")
+    assertThat(result.task(":bootJar")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
+
+    val file = findJar(projectDir, projectDetails.projectName)
+    val jarFile = JarFile(file)
+    assertThat(jarFile.manifest.mainAttributes.getValue("Implementation-Version")).isEqualTo(LocalDate.now().format(DateTimeFormatter.ISO_DATE))
+    assertThat(jarFile.manifest.mainAttributes.getValue("Implementation-Title")).isEqualTo(projectDetails.projectName)
+  }
+
+  @ParameterizedTest
+  @MethodSource("defaultProjectDetails")
+  fun `The gradle version dependency dependencyUpdates task is available`(projectDetails: ProjectDetails) {
+    makeProject(projectDetails)
+
+    val result = buildProject(projectDir, "dependencyUpdates", "-m")
+    assertThat(result.output)
+        .contains(":dependencyUpdates SKIPPED")
+        .contains("SUCCESSFUL")
+  }
+
+  @ParameterizedTest
+  @MethodSource("projectDetailsWithJunit4Tests")
+  fun `Junit 4 tests will not even compile`(projectDetails: ProjectDetails) {
+    makeProject(projectDetails)
+
+    buildProjectAndFail(projectDir, "compileTest")
+  }
+
+}
+
+private fun javaJunit4Test(): String {
+  return """
+      package uk.gov.justice.digital.hmpps.app;
+      
+      import org.junit.Test;
+
+      import static org.assertj.core.api.Assertions.assertThat;
+
+      public class ApplicationTest {
+          @Test
+          public void aTest() {
+              assertThat("anything").isEqualTo("anything");
+          }
+      }
+  """.trimIndent()
+}
+
+private fun kotlinJunit4Test(): String {
+  return """
+      package uk.gov.justice.digital.hmpps.app
+      
+      import org.assertj.core.api.Assertions.assertThat
+      import org.junit.Test
+      
+      class ApplicationTest {
+          @Test
+          fun `A Test`() {
+              assertThat("anything").isEqualTo("anything")
+          }
+      }
+  """.trimIndent()
+}
